Moved map-notify-edit component into separate folder
