+++ client/src/app/pages/dashboard/maps/map-info/map-info-notify-edit/map-info-notify-edit.component.html
@@ -1,0 +1,30 @@
+<nb-card>
+    <nb-card-header>
+        <h6>Edit Notification Settings</h6>
+    </nb-card-header>
+    <nb-card-body>
+        <div class="mb-3">Notify me when a</div>
+        <ul class="m-0 list-unstyled">
+            <li>
+                <nb-checkbox [(ngModel)]="checkboxFlags.PB.checked">
+                    personal best is set
+                </nb-checkbox>
+            </li>
+            <li>
+                <nb-checkbox [(ngModel)]="checkboxFlags.WR.checked">
+                    world record is set
+                </nb-checkbox>
+            </li>
+        </ul>
+    </nb-card-body>
+    <nb-card-footer>
+        <div class="row no-gutters">
+            <div class="col-6 pr-1">
+                <button class="btn btn-danger w-100" (click)="close()">Close</button>
+            </div>
+            <div class="col-6 pl-1">
+                <button class="btn btn-primary w-100" (click)="submit()">Submit</button>
+            </div>
+        </div>
+    </nb-card-footer>
+</nb-card>
