+++ client/src/app/pages/dashboard/maps/map-info/map-info-notify-edit/map-info-notify-edit.component.scss
@@ -1,0 +1,3 @@
+nb-card {
+  min-width: 300px;
+}
