+++ client/src/app/pages/dashboard/maps/map.module.ts
@@ -10,7 +10,8 @@ import {GallerizeModule} from '@ngx-gallery/gallerize';
 import {ViewMapsComponent} from './view-maps/view-maps.component';
 import {MapUploadFormComponent} from './upload-form/map-upload-form.component';
 import {UploadStatusComponent} from './upload-status/upload-status.component';
-import {MapInfoComponent, MapNotifyEditComponent} from './map-info/map-info.component';
+import {MapInfoComponent} from './map-info/map-info.component';
+import {MapNotifyEditComponent} from './map-info/map-info-notify-edit/map-info-notify-edit.component';
 import {MarkdownModule} from 'ngx-markdown';
 import {MapLibraryComponent} from './map-library/map-library.component';
 import {NbDatepickerModule, NbListModule, NbSelectModule, NbStepperModule, NbAlertModule} from '@nebular/theme';
